+++ cargo-solid/src/solidity_contract.rs
@@ -1,19 +1,20 @@
 use super::SolidityField;
 use serde::Deserialize;
+use std::path::PathBuf;
 
 #[derive(Debug, Deserialize)]
 pub struct SolidityContract {
-    pub filename: String,
+    pub filename: PathBuf,
     pub contract: String,
     pub fields: Vec<SolidityField>,
 }
 
 impl SolidityContract {
-    pub fn functions(&self) -> String {
+    pub fn functions(&self, nightly: bool) -> String {
         let functions = self
             .fields
             .iter()
-            .map(|field| field.to_rust_function())
+            .map(|field| field.to_rust_function(nightly))
             .collect::<Vec<_>>()
             .join("\n");
 
@@ -30,14 +31,14 @@ impl {contract} {{
         )
     }
 
-    pub fn types(&self) -> String {
+    pub fn types(&self, nightly: bool) -> String {
         format!(
             "\
 #[allow(unused_imports)]
 use solid::{{derive::Encode, Encode}};\n\n{}",
             self.fields
                 .iter()
-                .map(|field| field.get_output_type())
+                .map(|field| field.get_output_type(nightly))
                 .filter(Option::is_some)
                 .map(Option::unwrap)
                 .collect::<Vec<_>>()