+++ examples/cargo-solid-example/src/main.rs
@@ -1,3 +1,4 @@
+mod generated;
 mod stateful;
 
 fn main() {