+++ solid-core/src/bytesfix/mod.rs
@@ -3,3 +3,6 @@ pub mod nightly;
 pub mod stable;
 
 pub use stable::*;
+
+#[cfg(feature = "nightly")]
+pub use nightly::BytesFix;